requirements.txt
fastapi
uvicorn
tensorflow
numpy
pydantic
fastapi[standard]
pillow
requests
python-dotenv
google-cloud-storage